Move Complex viewer out of the xrefs and into a new section with tabs dedicated for the external viewers
